Update entity, CharacterServiceImpl ,GameServerServiceImpl

+++ src/main/java/com/example/user_web_service/entity/CharacterAttribute.java
@@ -17,7 +17,7 @@ public class CharacterAttribute {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Long value;
+    private float value;
 
     @ManyToOne
     @JsonIgnore
